add cache for common objects
