+++ pkg/monitor/cluster/clusterversion.go
@@ -7,14 +7,10 @@ import (
 	"context"
 
 	configv1 "github.com/openshift/api/config/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func (mon *Monitor) emitClusterVersion(ctx context.Context) error {
-	cv, err := mon.configcli.ConfigV1().ClusterVersions().Get("version", metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
+	cv := mon.cache.clusterVersion
 
 	desiredVersion := cv.Status.Desired.Version
 	if cv.Spec.DesiredUpdate != nil &&
