+++ pkg/monitor/cluster/podconditions_test.go
@@ -4,6 +4,7 @@ package cluster
 // Licensed under the Apache License 2.0.
 
 import (
+	"context"
 	"testing"
 
 	"github.com/golang/mock/gomock"
@@ -15,6 +16,8 @@ import (
 )
 
 func TestEmitPodConditions(t *testing.T) {
+	ctx := context.Background()
+
 	cli := fake.NewSimpleClientset(
 		&corev1.Pod{ // metrics expected
 			ObjectMeta: metav1.ObjectMeta{
@@ -57,6 +60,7 @@ func TestEmitPodConditions(t *testing.T) {
 		cli: cli,
 		m:   m,
 	}
+	mon.cache.podList, _ = cli.CoreV1().Pods("").List(metav1.ListOptions{})
 
 	m.EXPECT().EmitGauge("pods.conditions", int64(1), map[string]string{
 		"name":      "name",
@@ -83,54 +87,9 @@ func TestEmitPodConditions(t *testing.T) {
 		"type":      "Ready",
 	})
 
-	ps, _ := cli.CoreV1().Pods("").List(metav1.ListOptions{})
-	err := mon._emitPodConditions(ps)
+	err := mon.emitPodConditions(ctx)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-}
-
-func TestEmitPodContainersConditions(t *testing.T) {
-	cli := fake.NewSimpleClientset(
-		&corev1.Pod{ // metrics expected
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "name",
-				Namespace: "openshift",
-			},
-			Status: corev1.PodStatus{
-				ContainerStatuses: []corev1.ContainerStatus{
-					{
-						State: corev1.ContainerState{
-							Waiting: &corev1.ContainerStateWaiting{
-								Reason: "ImagePullBackOff",
-							},
-						},
-					},
-				},
-			},
-		},
-	)
-
-	controller := gomock.NewController(t)
-	defer controller.Finish()
-
-	m := mock_metrics.NewMockInterface(controller)
-
-	mon := &Monitor{
-		cli: cli,
-		m:   m,
-	}
-
-	m.EXPECT().EmitGauge("pods.containers.conditions", int64(1), map[string]string{
-		"name":      "name",
-		"namespace": "openshift",
-		"reason":    "ImagePullBackOff",
-	})
-
-	ps, _ := cli.CoreV1().Pods("").List(metav1.ListOptions{})
-	err := mon._emitPodContainersConditions(ps)
-	if err != nil {
-		t.Fatal(err)
-	}
 }
