+++ pkg/monitor/cluster/podcontainerconditions_test.go
@@ -1,0 +1,62 @@
+package cluster
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the Apache License 2.0.
+
+import (
+	"context"
+	"testing"
+
+	"github.com/golang/mock/gomock"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes/fake"
+
+	mock_metrics "github.com/Azure/ARO-RP/pkg/util/mocks/metrics"
+)
+
+func TestEmitPodContainersConditions(t *testing.T) {
+	ctx := context.Background()
+
+	cli := fake.NewSimpleClientset(
+		&corev1.Pod{ // metrics expected
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "name",
+				Namespace: "openshift",
+			},
+			Status: corev1.PodStatus{
+				ContainerStatuses: []corev1.ContainerStatus{
+					{
+						State: corev1.ContainerState{
+							Waiting: &corev1.ContainerStateWaiting{
+								Reason: "ImagePullBackOff",
+							},
+						},
+					},
+				},
+			},
+		},
+	)
+
+	controller := gomock.NewController(t)
+	defer controller.Finish()
+
+	m := mock_metrics.NewMockInterface(controller)
+
+	mon := &Monitor{
+		cli: cli,
+		m:   m,
+	}
+	mon.cache.podList, _ = cli.CoreV1().Pods("").List(metav1.ListOptions{})
+
+	m.EXPECT().EmitGauge("pods.containers.conditions", int64(1), map[string]string{
+		"name":      "name",
+		"namespace": "openshift",
+		"reason":    "ImagePullBackOff",
+	})
+
+	err := mon.emitPodContainersConditions(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+}
